AtomisticQoIs

Documentation for AtomisticQoIs.

AtomisticQoIs.AtomisticQoIs

AtomisticQoIs.GaussQuadrature

AtomisticQoIs.GibbsQoI

AtomisticQoIs.HMC

AtomisticQoIs.ISMC

AtomisticQoIs.ISMCMC

AtomisticQoIs.ISMixSamples

AtomisticQoIs.ISSamples

AtomisticQoIs.Integrator

AtomisticQoIs.MALA

AtomisticQoIs.MCMC

AtomisticQoIs.MCSamples

AtomisticQoIs.MonteCarlo

AtomisticQoIs.NUTS

AtomisticQoIs.QoI

AtomisticQoIs.EffSampleSize

AtomisticQoIs.EffSampleSize

AtomisticQoIs.ISWeightDiagnostic

AtomisticQoIs.ISWeightVariance

AtomisticQoIs.MCSE

AtomisticQoIs.MCSEbm

AtomisticQoIs.MCSEobm

AtomisticQoIs.compute_metrics

AtomisticQoIs.expectation_is_stable

AtomisticQoIs.gaussquad

AtomisticQoIs.gaussquad_2D

AtomisticQoIs.grad_expectation

AtomisticQoIs.hasapproxnormconst

AtomisticQoIs.hasupdf

AtomisticQoIs.sample

AtomisticQoIs.sample

AtomisticQoIs.sample

AtomisticQoIs.AtomisticQoIs — Module

AtomisticQoIs

Author: Joanna Zou <jjzou@mit.edu> Version: 0.1.0 Year: 2024 Notes: A Julia library for computing expectation-valued quantities of interest (QoIs) from molecular dynamics simulation.

AtomisticQoIs.GaussQuadrature — Type

struct GaussQuadrature <: QuadIntegrator

Contains parameters for 1D or 2D Gauss quadrature integration.

Arguments

ξ :: Union{Vector{<:Real}, Vector{<:Vector}} : quadrature points

w :: Union{Vector{<:Real}, Vector{<:Vector}} : quadrature weights

d :: Real : dimension

AtomisticQoIs.GibbsQoI — Type

@kwdef mutable struct GibbsQoI <: QoI

Defines the struct for computing the quantity of interest (QoI) as the expected value of a function h over Gibbs measure p given input parameters θ, e. g. E_p[h(θ)]

Arguments

h :: Function : function to evaluate expectation, receiving inputs (x,θ)

p :: Gibbs : Gibbs (invariant) distribution with θ == nothing (undefined parameters)

∇h :: Union{Function, Nothing} = nothing : gradient of h (if nothing, gradient is computed using AutoDiff)

AtomisticQoIs.HMC — Type

struct HMC <: Sampler

Defines the struct with parameters of the Hamiltonian Monte Carlo algorithm for sampling.

Arguments

L :: Int : time length of integrating Hamiltonian's equations

ϵ :: Real : step size

AtomisticQoIs.ISMC — Type

struct ISMC <: ISIntegrator

Contains parameters for Importance Sampling (IS) integration using MC samples from a biasing distribution. This method is implemented when the biasing distribution g can be analytically sampled using rand().

Arguments

g :: Distribution : biasing distribution

n :: Int : number of samples

AtomisticQoIs.ISMCMC — Type

struct ISMCMC <: ISIntegrator

Contains parameters for Importance Sampling (IS) integration using MCMC samples from a biasing distribution. This method is implemented when the biasing distribution g cannot be analytically sampled.

Arguments

g :: Distribution : biasing distribution

n :: Int : number of samples

sampler :: Sampler : type of sampler (see Sampler)

ρ0 :: Distribution : prior distribution of the state

AtomisticQoIs.ISMixSamples — Type

struct ISMixSamples <: ISIntegrator

Contains parameters for Importance Sampling (IS) integration using samples from a mixture biasing distribution. The mixture weights are computed based on a kernel distance metric of the parameter with respect to parameters of the mixture component distributions. This method is implemented with the user providing samples from each component mixture distribution.

Arguments

g :: MixtureModel : mixture biasing distribution

n :: Int : number of samples

knl :: Kernel : kernel function to compute mixture weights

xsamp :: Vector : Vector of sample sets from each component distribution

normint :: Integrator : Integrator for the approximating the normalizing constant of each component distribution

AtomisticQoIs.ISSamples — Type

struct ISSamples <: ISIntegrator

Contains pre-defined Monte Carlo samples from the biasing distribution to evaluate in integration. This method is implemented with the user providing samples from the biasing distribution.

Arguments

g :: Distribution : biasing distribution

xsamp :: Vector : fixed set of samples

normint :: Union{Integrator, Nothing} : integrator for computing normalizing constant of biasing distribution (required for mixture models)

AtomisticQoIs.Integrator — Type

Integrator

A struct of abstract type Integrator computes the expectation of a function h(x, θ) with respect to an invariant measure p(x, θ).

AtomisticQoIs.MALA — Type

struct MALA <: Sampler

Defines the struct with parameters of the Metropolis-Adjusted Langevin Algorithm (MALA) for sampling.

Arguments

L :: Int : time length between evaluating accept/reject criterion

ϵ :: Real : step size

AtomisticQoIs.MCMC — Type

struct MCMC <: MCIntegrator

Contains parameters for Monte Carlo (MC) integration using MCMC samples. This method is implemented when the distribution cannot be analytically sampled.

Arguments

n :: Int : number of samples

sampler :: Sampler : type of sampler (see Sampler)

ρ0 :: Distribution : prior distribution of the state

AtomisticQoIs.MCSamples — Type

struct MCSamples <: MCIntegrator

Contains pre-defined Monte Carlo samples to evaluate in integration. This method is implemented with the user providing samples from the distribution.

Arguments

xsamp :: Vector : fixed set of samples

AtomisticQoIs.MonteCarlo — Type

struct MonteCarlo <: MCIntegrator

Contains parameters for Monte Carlo (MC) integration using MC samples. This method may only be implemented when the distribution can be analytically sampled using rand().

Arguments

n :: Int : number of samples

AtomisticQoIs.NUTS — Type

struct NUTS <: Sampler

Defines the struct with parameters of the No-U-Turn-Sampler algorithm for sampling.

Arguments

ϵ :: Union{<:Real, Nothing} : step size

AtomisticQoIs.QoI — Type

QoI

A struct of abstract type QoI computes the expectation of a function h(x, θ) with respect to an invariant measure p(x, θ).

AtomisticQoIs.EffSampleSize — Method

function EffSampleSize(q::GibbsQoI, xsamp::Vector)

Computes the effective sample size of a set of nMC samples from an MCMC chain.

Arguments

q :: GibbsQoI : Gibbs QoI struct

xsamp :: Vector : nMC-vector of x-samples

Outputs

ESS :: Float64 : effective sample size (where 1 < ESS < nMC)

AtomisticQoIs.EffSampleSize — Method

function EffSampleSize(hsamp::Vector)

Computes the effective sample size of a set of nMC samples from an MCMC chain.

Arguments

hsamp :: Vector : nMC-vector of d-dimensional function evaluation h

Outputs

ESS :: Float64 : effective sample size (where 1 < ESS < nMC)

AtomisticQoIs.ISWeightDiagnostic — Method

function ISWeightDiagnostic(wsamp::Vector)

Returns a diagnostic value for importance sampling weights using an unnormalized biasing distribution. The importance sampling estimate is reliable when the diagnostic is less than 5.

Arguments

wsamp :: Vector : nMC-vector of 1-dim. weights

Outputs

diagnostic :: Float64 : diagnostic value (want < 5)

AtomisticQoIs.ISWeightVariance — Method

function ISWeightVariance(wsamp::Vector)

Returns the variance of importance sampling weights.

Arguments

wsamp :: Vector : nMC-vector of weights

Outputs

var(wsamp) :: Float64 : variance of weights

AtomisticQoIs.MCSE — Method

function MCSE(q::GibbsQoI, xsamp::Vector)

Estimates the Monte Carlo standard error (se = std(h(x))/√n) using an empirical estimate of standard deviation.

Arguments

q :: GibbsQoI : Gibbs QoI struct

xsamp :: Vector : vector of samples

Outputs

se :: Float64 : Monte Carlo standard error

AtomisticQoIs.MCSEbm — Method

function MCSEbm(q::GibbsQoI, xsamp::Vector)

Estimates the Monte Carlo standard error (se = std(h(x))/√n) using the batch means (BM) method.

Arguments

q :: GibbsQoI : Gibbs QoI struct

xsamp :: Vector : vector of samples

Outputs

se :: Float64 : Monte Carlo standard error

AtomisticQoIs.MCSEobm — Method

function MCSEbm(q::GibbsQoI, xsamp::Vector)

Estimates the Monte Carlo standard error (se = std(h(x))/√n) using the overlapping batch means (OBM) method.

Arguments

q :: GibbsQoI : Gibbs QoI struct

xsamp :: Vector : vector of samples

Outputs

se :: Float64 : Monte Carlo standard error

AtomisticQoIs.compute_metrics — Method

function compute_metrics(θ::Vector, h::Vector, w::Vector)

Computes importance sampling diagnostic metrics, including variance of IS weights ("wvar"), ESS of IS weights ("wESS"), and diagnostic measure ("wdiag").

Arguments

θ :: Vector : vector of parameters

h :: Vector : evaluations of summand function h(x) at samples

w :: Vector : evaluations of importance sampling weights at samples

Outputs

metrics :: Dict{String, Vector} : dictionary of importance sampling diagnostics

AtomisticQoIs.expectation_is_stable — Method

function expectationisstable(xsamp::Vector, ϕ::Function, f::Gibbs, g::Distribution; normint=nothing)

Helper function for computing stable importance sampling weights (using log formulation).

AtomisticQoIs.gaussquad — Method

function gaussquad(nquad::Integer, gaussbasis::Function, args...; limits::Vector=[-1,1])

Computes 1D Gauss quadrature points with a change of domain to limits=[ll, ul].

Arguments

nquad::Integer : number of quadrature points

gaussbasis::Function : polynomial basis function from FastGaussQuadrature (gausslegendre, gaussjacobi, etc.)

args... : additional arguments for the gaussbasis() function

limits::Vector : defines lower and upper limits for rescaling

Outputs

ξ :: Vector{<:Real} : quadrature points ranging [ll, ul]

w :: Vector{<:Real} : quadrature weights

AtomisticQoIs.gaussquad_2D — Method

function gaussquad_2D(nquad::Integer, gaussbasis::Function, args...; limits::Vector=[-1,1])

Computes 2D (symmetric) Gauss quadrature points with a change of domain to limits=[ll, ul].

Arguments

nquad::Integer : number of quadrature points

gaussbasis::Function : polynomial basis function from FastGaussQuadrature (gausslegendre, gaussjacobi, etc.)

args... : additional arguments for the gaussbasis() function

limits::Vector : defines lower and upper limits for rescaling

Outputs

ξ :: Vector{<:Real} : quadrature points ranging [ll, ul]

w :: Vector{<:Real} : quadrature weights

AtomisticQoIs.grad_expectation — Method

function grad_expectation(θ::Union{Real, Vector{<:Real}}, qoi::GibbsQoI, integrator::Integrator; gradh::Union{Function, Nothing}=nothing)

Computes the gradient of the QoI with respect to the parameters θ, e. g. ∇θ E_p[h(θ)]. See expectation for more information.

Arguments

θ :: Union{Real, Vector{<:Real}} : parameters to evaluate expectation

qoi :: QoI : QoI object containing h and p

integrator :: Integrator : struct specifying method of integration

gradh :: Union{Function, Nothing} : gradient of qoi.h; if nothing, compute with ForwardDiff

Outputs

expec_estimator :: Vector{<:Real} : estimate of vector-valued gradient of the expectation

AtomisticQoIs.hasapproxnormconst — Method

function hasapproxnormconst(d::Distribution)

Returns 'true' if distribution d requires an approximation of its normalizing constant.

Arguments

d :: Distribution : distribution to check

Outputs

flag :: Bool : true or false

AtomisticQoIs.hasupdf — Method

function hasupdf(d::Distribution)

Returns 'true' if distribution d has an unnormalized pdf function (updf(), logupdf()).

Arguments

d :: Distribution : distribution to check

Outputs

flag :: Bool : true or false

AtomisticQoIs.sample — Method

function sample(lπ::Function, gradlπ::Function, sampler::NUTS, n::Int64, x0::Real)

Draw samples from target distribution π(x) using the No-U-Turn Sampler (NUTS) for scalar-valued support (x::Real).

Arguments

lπ :: Function : log likelihood of target π

gradlπ :: Function : gradient of log likelihood of π

sampler :: HMC : Sampler struct specifying sampling algorithm

n :: Int64 : number of samples

x0 :: Real : initial state

Outputs

samples :: Vector : vector of samples from π

AtomisticQoIs.sample — Method

function sample(lπ::Function, gradlπ::Function, sampler::NUTS, n::Int64, x0::Vector{<:Real})

Draw samples from target distribution π(x) using the No-U-Turn Sampler (NUTS) from the AdvancedHMC package. Assumes multi-dimensional support (x::Vector).

Arguments

lπ :: Function : log likelihood of target π

gradlπ :: Function : gradient of log likelihood of π

sampler :: HMC : Sampler struct specifying sampling algorithm

n :: Int64 : number of samples

x0 :: Vector{<:Real} : initial state

Outputs

samples::Vector{Vector} : vector of samples from π

AtomisticQoIs.sample — Method

function sample(lπ::Function, gradlπ::Function, sampler::Union{HMC,MALA}, n::Int64, x0::Real)

Draw samples from target distribution π(x) for scalar-valued support (x::Real).

Arguments

lπ :: Function : log likelihood of target π

gradlπ :: Function : gradient of log likelihood of π

sampler :: Union{HMC,MALA} : Sampler struct specifying sampling algorithm

n :: Int64 : number of samples

x0 :: Real : initial state

Outputs

samples :: Vector : vector of samples from π